backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
bcrypt==4.1.2